+++ src/main/java/com/company/businessprocess/customer/CustomerServiceImpl.java
@@ -3,10 +3,12 @@ package com.company.businessprocess.customer;
 import com.company.businessprocess.dto.request.CustomerRequest;
 import com.company.businessprocess.dto.response.CustomerResponse;
 import com.company.businessprocess.entity.CustomerEntity;
+import com.company.businessprocess.entity.ProductEntity;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -41,6 +43,9 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public void deleteCustomer(Integer id) {
-        customerRepository.deleteById(id);
+        Optional<CustomerEntity> optionalCustomerEntity = customerRepository.findById(id);
+        if (optionalCustomerEntity.isPresent()) {
+            customerRepository.deleteById(id);
+        }
     }
 }
